requirements.txt
flask==2.3.3
jinja2==3.1.2
selenium==4.15.2
playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1